+++ src/layout/main/index.vue
@@ -1,19 +1,19 @@
 <!--
  * @Author: jack-pearson
  * @Date: 2021-12-07 20:37:19
- * @LastEditTime: 2022-01-11 16:38:04
+ * @LastEditTime: 2022-01-12 13:23:40
  * @LastEditors: jack-pearson
  * @FilePath: /yh-vue3-admin/src/layout/main/index.vue
  * @Description: 
 -->
 <template>
   <el-container>
-    <ParentView />
+    <LayoutParentView />
   </el-container>
 </template>
 
 <script setup lang="ts">
-import ParentView from "@/components/parentView/index.vue";
+import LayoutParentView from "@/layout/router-view/index.vue";
 </script>
 
 <style scoped lang="scss"></style>